move contract interaction page
