+++ packages/plugin-ethereum/client/src/components/contract.json
